Fix error add int() conversion

+++ median.py
@@ -13,6 +13,6 @@ print(numbers)
 
 size = len(numbers)
 if size%2 == 1: 
-    print(numbers[(size-1)/2])
+    print(numbers[int((size-1)/2)])
 else: 
-    print((numbers[size/2]+numbers[(size/2)-1])/2)
+    print((numbers[int(size/2)]+numbers[int((size/2)-1)])/2)
